catalog-backend: move github util to own folder
