+++ plugins/catalog-backend/src/ingestion/processors/index.ts
@@ -16,11 +16,6 @@
 
 import * as results from './results';
 
-export { results };
-export * from './types';
-
-export { parseEntityYaml } from './util/parse';
-
 export { AnnotateLocationEntityProcessor } from './AnnotateLocationEntityProcessor';
 export { BuiltinKindsEntityProcessor } from './BuiltinKindsEntityProcessor';
 export { CodeOwnersProcessor } from './CodeOwnersProcessor';
@@ -32,4 +27,7 @@ export { MicrosoftGraphOrgReaderProcessor } from './MicrosoftGraphOrgReaderProce
 export { PlaceholderProcessor } from './PlaceholderProcessor';
 export type { PlaceholderResolver } from './PlaceholderProcessor';
 export { StaticLocationProcessor } from './StaticLocationProcessor';
+export * from './types';
 export { UrlReaderProcessor } from './UrlReaderProcessor';
+export { parseEntityYaml } from './util/parse';
+export { results };
